Fix service startup on RHEL platforms
 link the missing libdnet.1 file to libdnet.so.101 so is actually provided by the libd package
